Split core.py into smaller files (#363)
