+++ pytype-conf.cfg
@@ -11,12 +11,8 @@ exclude =
     src/xpk/commands
     src/xpk/core/tests
     src/xpk/core/__init__.py
-    src/xpk/core/app_profile.py
-    src/xpk/core/blueprint.py
     src/xpk/core/cluster_private.py
     src/xpk/core/commands.py
-    src/xpk/core/core.py
-    src/xpk/core/job_template.py
     src/xpk/core/kjob.py
     src/xpk/core/kueue.py
     src/xpk/core/nap.py
