+++ src/xpk/commands/run.py
@@ -16,12 +16,12 @@ limitations under the License.
 
 from argparse import Namespace
 
+from ..core.commands import run_command_with_full_controls
+from ..core.gcloud_context import add_zone_and_project
+from ..core.kjob import AppProfileDefaults
 from ..core.kueue import LOCAL_QUEUE_NAME
 from ..utils.console import xpk_exit, xpk_print
 from .common import set_cluster_command
-from ..core.core import add_zone_and_project
-from ..core.kjob import AppProfileDefaults
-from ..core.commands import run_command_with_full_controls
 from .kind import set_local_cluster_command
