+++ src/xpk/core/config.py
@@ -14,24 +14,35 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-import ruamel.yaml
+import importlib.metadata as importlib_metadata
 import os
+import re
+
+import ruamel.yaml
 
 from ..utils import file
 from ..utils.console import xpk_print
+from .system_characteristics import AcceleratorType, SystemCharacteristics
 
+# This is the version for XPK PyPI package
+__version__ = importlib_metadata.version('xpk')
+XPK_CURRENT_VERSION = __version__.split('+')[0]
+XPK_CONFIG_FILE = os.path.expanduser('~/.config/xpk/config.yaml')
+
+CONFIGS_KEY = 'configs'
 CFG_BUCKET_KEY = 'cluster-state-gcs-bucket'
 CLUSTER_NAME_KEY = 'cluster-name'
 PROJECT_KEY = 'project-id'
 ZONE_KEY = 'zone'
-CONFIGS_KEY = 'configs'
-XPK_CONFIG_FILE = os.path.expanduser('~/.config/xpk/config.yaml')
-default_keys = [
+DEFAULT_KEYS = [
     CFG_BUCKET_KEY,
     CLUSTER_NAME_KEY,
     PROJECT_KEY,
     ZONE_KEY,
 ]
+VERTEX_TENSORBOARD_FEATURE_FLAG = XPK_CURRENT_VERSION >= '0.4.0'
+GCS_FUSE_ANNOTATION = 'gke-gcsfuse/volumes: "true"'
+
 
 yaml = ruamel.yaml.YAML()
 
@@ -41,7 +52,7 @@ class XpkConfig:
 
   def __init__(self, custom_config_file: str = XPK_CONFIG_FILE) -> None:
     self._config = custom_config_file
-    self._allowed_keys = default_keys
+    self._allowed_keys = DEFAULT_KEYS
 
   def _open_configs(self) -> dict | None:
     dir_path = '/'.join(self._config.split('/')[:-1])
@@ -91,3 +102,65 @@ class XpkConfig:
       return None
     val: dict[str, str] = config_yaml[CONFIGS_KEY]
     return val
+
+
+def parse_env_config(args, tensorboard_config, system: SystemCharacteristics):
+  """Parses the environment configurations to the jobset config.
+
+  Args:
+    args: user provided arguments for running the command.
+    tensorboard_config: configuration of Vertex Tensorboard.
+    system: system characteristics.
+  """
+  env = {}
+
+  env_pat = re.compile(r'(^[a-zA-Z_][a-zA-Z0-9_]*?)(?:=(.*))?$', re.M)
+  if args.env_file:
+    print('Setting container environment from', args.env_file)
+    with open(file=args.env_file, mode='r', encoding='utf-8') as f:
+      for match in env_pat.finditer(f.read()):
+        variable = match.group(1)
+        if match.group(2) is not None:
+          env[variable] = match.group(2)
+        else:
+          assert variable in os.environ, (
+              f'Variable {variable} is not set in the current '
+              'environment, a value must be specified.'
+          )
+          env[variable] = os.environ[variable]
+  if args.env:
+    for var in args.env:
+      match = env_pat.match(var)
+      assert match and match.group(2) is not None, (
+          'Invalid environment variable, format must be '
+          f'`--env VARIABLE=value`: {var}'
+      )
+      variable = match.group(1)
+      env[variable] = match.group(2)
+
+  if not args.use_pathways:
+    if args.debug_dump_gcs:
+      if 'XLA_FLAGS' in env:
+        raise ValueError(
+            'Conflict: XLA_FLAGS defined in both --debug_dump_gcs '
+            'and environment file. Please choose one way to define '
+            'XLA_FLAGS.'
+        )
+      env['XLA_FLAGS'] = '--xla_dump_to=/tmp/xla_dump/'
+
+    if tensorboard_config:
+      env['UPLOAD_DATA_TO_TENSORBOARD'] = True
+      for key, value in tensorboard_config.items():
+        env[key.upper()] = value
+
+  if system.accelerator_type == AcceleratorType['GPU']:
+    # For GPUs, it has two more spaces ahead of name and value respectively
+    env_format = '''
+                  - name: {key}
+                    value: "{value}"'''
+  else:
+    env_format = '''
+                - name: {key}
+                  value: "{value}"'''
+
+  args.env = ''.join(env_format.format(key=k, value=v) for k, v in env.items())
